Fix for constant authentication issues
When access token expired, auth middleware successfully grabbed new
tokens, but failed to update the request for handling. This led to failed
requests, and the need to reauth.

Now credentials are set in the request context and updated in the response,
preventing double token reauth with the same refresh token.